Add ActionHandler support for conditional executions
By adding a "when" clause on either an action or event, the action(s)
will only be executed if that expression is truthfull.
The expression itself is a Jinja2 expression which can evaulate against
the devices & their channel states.

To support per-eventtype conditions, a YAML change was introduced.
Instead of providing a list of actions as value of the event type,
a dict with 'actions'/'when' can be given given, where the 'actions' is the
list of actions.
The old syntax is still supported.